reordered test functions, removed old code
