+++ ark/utils/segmentation_utils_test.py
@@ -1,38 +1,9 @@
 import numpy as np
 import xarray as xr
-import os
-import pytest
-import tempfile
 
-import skimage.morphology as morph
-from skimage.morphology import erosion
 from skimage.measure import regionprops
 
 from ark.utils import segmentation_utils
-
-
-def _generate_deepcell_output(fov_num=2):
-    fovs = ["fov" + str(i) for i in range(fov_num)]
-    models = ["pixelwise_interior", "watershed_inner", "watershed_outer",
-              "fgbg_foreground", "pixelwise_sum"]
-    output = np.random.rand(len(fovs) * 50 * 50 * len(models))
-    output = output.reshape((len(fovs), 50, 50, len(models)))
-
-    output_xr = xr.DataArray(output, coords=[fovs, range(50), range(50), models],
-                             dims=["fovs", "rows", "cols", "models"])
-    return output_xr
-
-
-def _generate_channel_xr(fov_num=2, chan_num=5):
-    fovs = ["fov" + str(i) for i in range(fov_num)]
-    channels = ["channel" + str(i) for i in range(chan_num)]
-    output = np.random.randint(0, 20, len(fovs) * 50 * 50 * len(channels))
-    output = output.reshape((len(fovs), 50, 50, len(channels)))
-
-    output_xr = xr.DataArray(output, coords=[fovs, range(50), range(50), channels],
-                             dims=["fovs", "rows", "cols", "channels"])
-
-    return output_xr
 
 
 def _create_test_extraction_data():
@@ -60,129 +31,6 @@ def _create_test_extraction_data():
     channel_data[cell2, 4] = 10
 
     return cell_mask, channel_data
-
-
-
-
-
-def test_watershed_transform():
-    model_output = _generate_deepcell_output()
-    channel_data = _generate_channel_xr()
-
-    overlay_channels = [channel_data.channels.values[:2]]
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # test default settings
-        segmentation_utils.watershed_transform(model_output=model_output, channel_xr=channel_data,
-                                               output_dir=temp_dir,
-                                               overlay_channels=overlay_channels,
-                                               interior_threshold=0.5)
-
-        saved_output = xr.load_dataarray(os.path.join(temp_dir, 'segmentation_labels.xr'))
-
-        # ensure sequential labeling
-        for fov in range(saved_output.shape[0]):
-            cell_num = len(np.unique(saved_output[fov, :, :, 0]))
-            max_val = np.max(saved_output[fov, :, :, 0].values)
-            assert cell_num == max_val + 1  # background counts towards unique cells
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # test different networks settings
-        segmentation_utils.watershed_transform(model_output=model_output, channel_xr=channel_data,
-                                               output_dir=temp_dir,
-                                               overlay_channels=overlay_channels,
-                                               maxima_model="watershed_inner",
-                                               interior_model="watershed_outer")
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # test save_all tifs
-        segmentation_utils.watershed_transform(model_output=model_output, channel_xr=channel_data,
-                                               output_dir=temp_dir,
-                                               overlay_channels=overlay_channels,
-                                               save_tifs='all')
-        assert os.path.exists(os.path.join(temp_dir, 'fov1_interior_smoothed.tiff'))
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # only a subset
-        segmentation_utils.watershed_transform(model_output=model_output, channel_xr=channel_data,
-                                               output_dir=temp_dir,
-                                               overlay_channels=overlay_channels,
-                                               save_tifs='overlays')
-        assert not os.path.exists(os.path.join(temp_dir, 'fov1_interior_smoothed.tiff'))
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # test multiple different overlay_channels
-        overlay_channels = [channel_data.channels.values[3:4], channel_data.channels.values[1:4]]
-        segmentation_utils.watershed_transform(model_output=model_output, channel_xr=channel_data,
-                                               output_dir=temp_dir,
-                                               overlay_channels=overlay_channels)
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # bad arguments
-
-        with pytest.raises(ValueError):
-            # bad fov names
-            fovs = [model_output.fovs.values[0], 'bad_name']
-            segmentation_utils.watershed_transform(model_output=model_output,
-                                                   channel_xr=channel_data,
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=overlay_channels,
-                                                   fovs=fovs)
-
-        with pytest.raises(ValueError):
-            # incomplete channel data
-            segmentation_utils.watershed_transform(model_output=model_output,
-                                                   channel_xr=channel_data[1:, ...],
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=overlay_channels)
-
-        with pytest.raises(ValueError):
-            # bad overlay channels
-            bad_overlay_channels = [[channel_data.channels.values[0], 'bad_channel']]
-            segmentation_utils.watershed_transform(model_output=model_output,
-                                                   channel_xr=channel_data,
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=bad_overlay_channels)
-
-        with pytest.raises(ValueError):
-            # bad directory name
-            segmentation_utils.watershed_transform(model_output=model_output,
-                                                   channel_xr=channel_data,
-                                                   output_dir='bad_output_dir',
-                                                   overlay_channels=overlay_channels)
-
-        with pytest.raises(ValueError):
-            # bad maxima model name
-            segmentation_utils.watershed_transform(model_output=model_output,
-                                                   channel_xr=channel_data,
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=overlay_channels,
-                                                   maxima_model='bad_maxima_model')
-
-        with pytest.raises(ValueError):
-            # maxima model not included
-            missing_pixel_model = model_output[..., 2:]
-            segmentation_utils.watershed_transform(model_output=missing_pixel_model,
-                                                   channel_xr=channel_data,
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=overlay_channels)
-
-        with pytest.raises(ValueError):
-            # bad interior model name
-            segmentation_utils.watershed_transform(model_output=model_output,
-                                                   channel_xr=channel_data,
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=overlay_channels,
-                                                   interior_model='bad_interior_model')
-
-        with pytest.raises(ValueError):
-            # maxima model not included
-            missing_interior_model = model_output[..., 1:]
-            segmentation_utils.watershed_transform(model_output=missing_interior_model,
-                                                   maxima_model='watershed_inner',
-                                                   channel_xr=channel_data,
-                                                   output_dir=temp_dir,
-                                                   overlay_channels=overlay_channels)
 
 
 def test_find_nuclear_mask_id():
@@ -316,54 +164,3 @@ def test_transform_expression_matrix_multiple_compartments():
     for cell in cell_data.cell_id:
         arcsinh_vals = np.arcsinh(cell_data.loc[:, cell, modified_cols].values)
         assert np.array_equal(arcsinh_data.loc[:, cell, modified_cols].values, arcsinh_vals)
-
-
-# TODO: The testing for this function can be improved. Repeated code, and lots of manual checks
-
-
-# TODO: refactor these tests to share code
-
-
-def test_generate_expression_matrix_multiple_compartments():
-    cell_mask, channel_data = _create_test_extraction_data()
-
-    # generate data for two fovs offset
-    cell_masks = np.zeros((2, 40, 40, 1), dtype="int16")
-    cell_masks[0, :, :, 0] = cell_mask
-    cell_masks[1, 5:, 5:, 0] = cell_mask[:-5, :-5]
-
-    channel_datas = np.zeros((2, 40, 40, 5), dtype="int16")
-    channel_datas[0, :, :, :] = channel_data
-    channel_datas[1, 5:, 5:, :] = channel_data[:-5, :-5]
-
-    # generate a second set of nuclear masks that are smaller than cell masks
-    nuc_masks = np.zeros_like(cell_masks)
-    nuc_masks[0, :, :, 0] = erosion(cell_masks[0, :, :, 0], selem=morph.disk(1))
-    nuc_masks[1, :, :, 0] = erosion(cell_masks[1, :, :, 0], selem=morph.disk(1))
-
-    # cell 2 in fov0 has no nucleus
-    nuc_masks[0, nuc_masks[0, :, :, 0] == 2, 0] = 0
-
-    unequal_masks = np.concatenate((cell_masks, nuc_masks), axis=-1)
-    coords = [["Point0", "Point1"], range(40), range(40), ['whole_cell', 'nuclear']]
-    dims = ['fovs', 'rows', 'cols', 'compartments']
-    segmentation_masks_unequal = xr.DataArray(unequal_masks, coords=coords, dims=dims)
-
-    channel_data = xr.DataArray(channel_datas,
-                                coords=[["Point0", "Point1"], range(40), range(40),
-                                        ["chan0", "chan1", "chan2", "chan3", "chan4"]],
-                                dims=["fovs", "rows", "cols", "channels"])
-
-    normalized, arcsinh = segmentation_utils.generate_expression_matrix(segmentation_masks_unequal,
-                                                                        channel_data,
-                                                                        nuclear_counts=True)
-
-    assert normalized.shape[0] == 7
-
-    assert np.all(normalized['chan0'] == np.repeat(1, len(normalized)))
-    assert np.all(normalized['chan1'] == np.repeat(5, len(normalized)))
-    assert np.all(normalized['chan2'] == normalized['chan2'])
-
-    # check that missing nucleus has size 0
-    index = np.logical_and(normalized['label'] == 2, normalized['fov'] == 'Point0')
-    assert normalized.loc[index, 'cell_size_nuclear'].values == 0
